# docs/swagger.yaml
basePath: /api
info:
  contact: {}
  description: API for managing groceries
  title: Grocery API
  version: "1.0"
paths:
  /CreateGrocery:
    post:
      consumes:
      - application/json
      description: Create a new grocery item with the provided data and image
      operationId: create-grocery
      parameters:
      - description: JSON data for the grocery item
        in: formData
        name: json-data
        required: true
        type: string
      - description: Image file for the grocery item
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: File uploaded successfully
          schema:
            type: string
        "400":
          description: 'Bad Request: Invalid JSON payload or missing required fields'
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Create a new grocery item
...
  /api/bulkUploadGroceryItems:
    post:
      consumes:
      - multipart/form-data
      description: Uploads multiple grocery items from a CSV or JSON file
      operationId: bulk-upload-grocery-items
      parameters:
      - description: CSV or JSON file containing grocery items
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: File URL sent successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 'Bad Request: Unsupported file type. Only CSV or JSON files
            are allowed'
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Bulk upload grocery items
swagger: "2.0"